requirements.txt
crewai==0.86.0
langchain==0.2.16
langchain-community==0.2.10
langchain-core==0.2.41
openai==1.45.0
tiktoken==0.7.0
beautifulsoup4==4.12.3
html5lib==1.1
lxml==5.2.2
requests==2.32.3
ratelimit==2.2.1
tenacity==9.0.0
rapidfuzz==3.9.7
regex==2024.9.11
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.5
loguru==0.7.2
pyjwt==2.9.0
httpx==0.27.2
duckdb==1.1.1
uvicorn==0.30.6
typer==0.12.5
tabulate==0.9.0
orjson==3.10.7